requirements.txt
albumentations==1.3.0
easydict
lightning==2.0.2
matplotlib
notebook
numpy
opencv-python==4.7.*
pandas
pytorch-lightning==1.8.6
pylsd-nova
PyYAML
tensorboard
torch==1.13.1
torchvision==0.14.1
tqdm